Function rustc_mir_dataflow::drop_flag_effects::on_all_inactive_variants

pub(crate) fn on_all_inactive_variants<'tcx>(
    move_data: &MoveData<'tcx>,
    enum_place: Place<'tcx>,
    active_variant: VariantIdx,
    handle_inactive_variant: impl FnMut(MovePathIndex)
)

Calls handle_inactive_variant for each descendant move path of enum_place that contains a Downcast to a variant besides the active_variant.

NOTE: If there are no move paths corresponding to an inactive variant, handle_inactive_variant will not be called for that variant.