rustc_mir_dataflow::framework::cursor

Enum CowMut

enum CowMut<'a, T> {
    BorrowedMut(&'a mut T),
    Owned(T),
}

Some ResultsCursors want to own an Analysis, and some want to borrow an Analysis, either mutable or immutably. This type allows all of the above. Itโs similar to Cow, but Cow doesnโt allow mutable borrowing.

Variants

BorrowedMut(&'a mut T)

Owned(T)