rustc_mir_dataflow::framework::engine
Function write_graphviz_results

fn write_graphviz_results<'tcx, A>(
    tcx: TyCtxt<'tcx>,
    body: &Body<'tcx>,
    results: Results<'tcx, A>,
    pass_name: Option<&'static str>,
) -> (Result<()>, Results<'tcx, A>)
where
    A: Analysis<'tcx>,
    A::Domain: DebugWithContext<A>,

Writes a DOT file containing the results of a dataflow analysis if the user requested it via rustc_mir attributes and -Z dump-mir-dataflow. The Result in and the Results out are the same.