Struct rustc_mir_dataflow::framework::engine::Results

pub struct Results<'tcx, A, E = IndexVec<BasicBlock, <A as AnalysisDomain<'tcx>>::Domain>>where
    A: Analysis<'tcx>,{
    pub analysis: A,
    pub(super) entry_sets: E,
    pub(super) _marker: PhantomData<&'tcx ()>,
}

A dataflow analysis that has converged to fixpoint.

Fields
analysis: Aentry_sets: E_marker: PhantomData<&'tcx ()>
Implementations

impl<'tcx, A, E> Results<'tcx, A, E>where A: Analysis<'tcx>, E: Borrow<IndexVec<BasicBlock, <A as AnalysisDomain<'tcx>>::Domain>>,

pub fn into_results_cursor<'mir>( self, body: &'mir Body<'tcx> ) -> ResultsCursor<'mir, 'tcx, A, Self>

Creates a ResultsCursor that can inspect these Results.

pub fn entry_set_for_block(&self, block: BasicBlock) -> &A::Domain

Gets the dataflow state for the given block.

pub fn visit_with<'mir>( &mut self, body: &'mir Body<'tcx>, blocks: impl IntoIterator<Item = BasicBlock>, vis: &mut impl ResultsVisitor<'mir, 'tcx, Self, FlowState = A::Domain> )

pub fn visit_reachable_with<'mir>( &mut self, body: &'mir Body<'tcx>, vis: &mut impl ResultsVisitor<'mir, 'tcx, Self, FlowState = A::Domain> )

impl<'tcx, A> Results<'tcx, A>where A: Analysis<'tcx>,

pub fn as_results_cursor<'a, 'mir>( &'a mut self, body: &'mir Body<'tcx> ) -> ResultsRefCursor<'a, 'mir, 'tcx, A>

Creates a ResultsCursor that can inspect these Results.

impl<'tcx, A> Results<'tcx, A>where A: Analysis<'tcx> + CloneAnalysis,

pub fn clone_analysis(&self) -> ResultsCloned<'_, 'tcx, A>

Creates a new Results type with a cloned Analysis and borrowed entry sets.

pub fn cloned_results_cursor<'mir>( &self, body: &'mir Body<'tcx> ) -> ResultsClonedCursor<'_, 'mir, 'tcx, A>

Creates a ResultsCursor that can inspect these Results.

impl<'res, 'tcx, A> Results<'tcx, A, &'res IndexVec<BasicBlock, <A as AnalysisDomain<'tcx>>::Domain>>where A: Analysis<'tcx> + CloneAnalysis,

pub fn reclone_analysis(&self) -> Self

Creates a new Results type with a cloned Analysis and borrowed entry sets.