Struct rustc_mir_dataflow::framework::graphviz::BlockFormatter

struct BlockFormatter<'a, 'tcx, A>where
    A: Analysis<'tcx>,{
    results: ResultsRefCursor<'a, 'a, 'tcx, A>,
    bg: Background,
    style: OutputStyle,
}

Fields
results: ResultsRefCursor<'a, 'a, 'tcx, A>bg: Backgroundstyle: OutputStyle
Implementations

impl<'a, 'tcx, A> BlockFormatter<'a, 'tcx, A>where
A: Analysis<'tcx>,
A::Domain: DebugWithContext<A>,

const HEADER_COLOR: &'static str = "#a0a0a0"

fn toggle_background(&mut self) -> Background

fn write_node_label(
&mut self,
w: &mut impl Write,
body: &'a Body<'tcx>,
block: BasicBlock
) -> Result<()>

fn write_block_header_simple(
&mut self,
w: &mut impl Write,
block: BasicBlock
) -> Result<()>

fn write_block_header_with_state_columns(
&mut self,
w: &mut impl Write,
block: BasicBlock,
state_column_names: &[&str]
) -> Result<()>

fn write_statements_and_terminator(
&mut self,
w: &mut impl Write,
body: &'a Body<'tcx>,
block: BasicBlock
) -> Result<()>

fn write_row<W: Write>(
&mut self,
w: &mut W,
i: &str,
mir: &str,
f: impl FnOnce(&mut Self, &mut W, &str) -> Result<()>
) -> Result<()>

Write a row with the given index and MIR, using the function argument to fill in the “STATE” column(s).

fn write_row_with_full_state(
&mut self,
w: &mut impl Write,
i: &str,
mir: &str
) -> Result<()>