Struct rustc_mir_dataflow::framework::lattice::Dual

pub struct Dual<T>(pub T);

The counterpart of a given semilattice T using the inverse order.

The dual of a join-semilattice is a meet-semilattice and vice versa. For example, the dual of a powerset has the empty set as its top element and the full set as its bottom element and uses set intersection as its join operator.

Tuple Fields
0: T