rustc_mir_dataflow::framework::results

Type Alias Results

pub type Results<D> = IndexVec<BasicBlock, D>;

The results of a dataflow analysis that has converged to fixpoint. It only holds the domain values at the entry of each basic block. Domain values in other parts of the block are recomputed on the fly by visitors (i.e. ResultsCursor, or ResultsVisitor impls).

Aliased Type

#[repr(transparent)]
pub struct Results<D> {
    pub raw: Vec<D>,
    /* private fields */
}

Fields
raw: Vec<D>
Layout

Note: Most layout information is completely unstable and may even differ between compilations. The only exception is types with certain repr(...) attributes. Please see the Rust Reference's “Type Layout” chapter for details on type layout guarantees.

Size: 24 bytes