rustc_mir_dataflow::impls

Module initialized

Structs

EverInitializedPlaces

EverInitializedPlaces tracks all places that might have ever been initialized upon reaching a particular point in the control flow for a function, without an intervening StorageDead.

MaybeInitializedPlaces

MaybeInitializedPlaces tracks all places that might be initialized upon reaching a particular point in the control flow for a function.

MaybePlacesSwitchIntData

MaybeUninitializedPlaces

MaybeUninitializedPlaces tracks all places that might be uninitialized upon reaching a particular point in the control flow for a function.

Type Aliases

EverInitializedPlacesDomain

There can be many more InitIndex than there are locals in a MIR body. We use a mixed bitset to avoid paying too high a memory footprint.

MaybeUninitializedPlacesDomain

There can be many more MovePathIndex than there are locals in a MIR body. We use a mixed bitset to avoid paying too high a memory footprint.