rustc_mir_dataflow::impls::initialized

Type Alias EverInitializedPlacesDomain

pub type EverInitializedPlacesDomain = MixedBitSet<InitIndex>;

There can be many more InitIndex than there are locals in a MIR body. We use a mixed bitset to avoid paying too high a memory footprint.

Aliased Type

pub enum EverInitializedPlacesDomain {
    Small(DenseBitSet<InitIndex>),
    Large(ChunkedBitSet<InitIndex>),
}

Variants

Small(DenseBitSet<InitIndex>)

Large(ChunkedBitSet<InitIndex>)

Layout

Note: Most layout information is completely unstable and may even differ between compilations. The only exception is types with certain repr(...) attributes. Please see the Rust Reference's “Type Layout” chapter for details on type layout guarantees.

Size: 40 bytes

Size for each variant:

Small: 32 bytes

Large: 24 bytes