Struct rustc_mir_dataflow::impls::liveness::MaybeTransitiveLiveLocals

pub struct MaybeTransitiveLiveLocals<'a> {
    always_live: &'a BitSet<Local>,
}

Like MaybeLiveLocals, but does not mark locals as live if they are used in a dead assignment.

This is basically written for dead store elimination and nothing else.

All of the caveats of MaybeLiveLocals apply.

Fields
always_live: &'a BitSet<Local>
Implementations

impl<'a> MaybeTransitiveLiveLocals<'a>

pub fn new(always_live: &'a BitSet<Local>) -> Self

The always_alive set is the set of locals to which all stores should unconditionally be considered live.

This should include at least all locals that are ever borrowed.