rustc_mir_dataflow::impls
Module storage_liveness

Structs

MaybeRequiresStorage

Dataflow analysis that determines whether each local requires storage at a given location; i.e. whether its storage can go away without being observed.

MaybeStorageDead

MaybeStorageLive

MoveVisitor 🔒

Functions

always_storage_live_locals

The set of locals in a MIR body that do not have StorageLive/StorageDead annotations.

Type Aliases

BorrowedLocalsResults 🔒