Struct rustc_mir_dataflow::impls::MaybeLiveLocals

pub struct MaybeLiveLocals;

A live-variable dataflow analysis.

This analysis considers references as being used only at the point of the borrow. In other words, this analysis does not track uses because of references that already exist. See this mir-dataflow test for an example. You almost never want to use this analysis without also looking at the results of MaybeBorrowedLocals.

Field-(in)sensitivity

As the name suggests, this analysis is field insensitive. If a projection of a variable x is assigned to (e.g. x.0 = 42), it does not “define” x as far as liveness is concerned. In fact, such an assignment is currently marked as a “use” of x in an attempt to be maximally conservative.