rustc_mir_dataflow::impls
Type Alias MaybeUninitializedPlacesDomain

pub type MaybeUninitializedPlacesDomain = MixedBitSet<MovePathIndex>;

There can be many more MovePathIndex than there are locals in a MIR body. We use a mixed bitset to avoid paying too high a memory footprint.

Aliased Type

enum MaybeUninitializedPlacesDomain {
    Small(BitSet<MovePathIndex>),
    Large(ChunkedBitSet<MovePathIndex>),
}

Variants

Small(BitSet<MovePathIndex>)

Large(ChunkedBitSet<MovePathIndex>)

Layout

Note: Most layout information is completely unstable and may even differ between compilations. The only exception is types with certain repr(...) attributes. Please see the Rust Reference's “Type Layout” chapter for details on type layout guarantees.

Size: 40 bytes

Size for each variant:

Small: 32 bytes

Large: 24 bytes