Struct rustc_mir_dataflow::move_paths::MovePathLookup

pub struct MovePathLookup {
    locals: IndexVec<Local, MovePathIndex>,
    projections: FxHashMap<(MovePathIndex, ProjectionElem<AbstractOperand, AbstractType>), MovePathIndex>,
}

Tables mapping from a place to its MovePathIndex.

Fields
locals: IndexVec<Local, MovePathIndex>projections: FxHashMap<(MovePathIndex, ProjectionElem<AbstractOperand, AbstractType>), MovePathIndex>

projections are made from a base-place and a projection elem. The base-place will have a unique MovePathIndex; we use the latter as the index into the outer vector (narrowing subsequent search so that it is solely relative to that base-place). For the remaining lookup, we map the projection elem to the associated MovePathIndex.

Implementations

impl MovePathLookup

pub fn find(&self, place: PlaceRef<'_>) -> LookupResult

pub fn find_local(&self, local: Local) -> MovePathIndex

pub fn iter_locals_enumerated(
&self
) -> impl DoubleEndedIterator<Item = (Local, MovePathIndex)> + ExactSizeIterator + '_

An enumerated iterator of locals and their associated MovePathIndexes.