rustc_mir_dataflow
Trait JoinSemiLattice

pub trait JoinSemiLattice: Eq {
    // Required method
    fn join(&mut self, other: &Self) -> bool;
}

A partially ordered set that has a least upper bound for any pair of elements in the set.

Required Methods

fn join(&mut self, other: &Self) -> bool

Computes the least upper bound of two elements, storing the result in self and returning true if self has changed.

The lattice join operator is abbreviated as ∨.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.