Type Alias rustc_mir_dataflow::ResultsCloned

pub type ResultsCloned<'res, 'tcx, A> = Results<'tcx, A, &'res IndexVec<BasicBlock, <A as AnalysisDomain<'tcx>>::Domain>>;

Results type with a cloned Analysis and borrowed entry sets.

Aliased Type

struct ResultsCloned<'res, 'tcx, A> {
    pub analysis: A,
    pub(super) entry_sets: &'res IndexVec<BasicBlock, <A as AnalysisDomain<'tcx>>::Domain>,
    pub(super) _marker: PhantomData<&'tcx ()>,
}

Fields
analysis: Aentry_sets: &'res IndexVec<BasicBlock, <A as AnalysisDomain<'tcx>>::Domain>_marker: PhantomData<&'tcx ()>
Implementations

impl<'tcx, A, E> Results<'tcx, A, E>where A: Analysis<'tcx>, E: Borrow<IndexVec<BasicBlock, <A as AnalysisDomain<'tcx>>::Domain>>,

pub fn into_results_cursor<'mir>( self, body: &'mir Body<'tcx> ) -> ResultsCursor<'mir, 'tcx, A, Self>

Creates a ResultsCursor that can inspect these Results.

pub fn entry_set_for_block(&self, block: BasicBlock) -> &A::Domain

Gets the dataflow state for the given block.

pub fn visit_with<'mir>( &mut self, body: &'mir Body<'tcx>, blocks: impl IntoIterator<Item = BasicBlock>, vis: &mut impl ResultsVisitor<'mir, 'tcx, Self, FlowState = A::Domain> )

pub fn visit_reachable_with<'mir>( &mut self, body: &'mir Body<'tcx>, vis: &mut impl ResultsVisitor<'mir, 'tcx, Self, FlowState = A::Domain> )

impl<'res, 'tcx, A> Results<'tcx, A, &'res IndexVec<BasicBlock, <A as AnalysisDomain<'tcx>>::Domain>>where A: Analysis<'tcx> + CloneAnalysis,

pub fn reclone_analysis(&self) -> Self

Creates a new Results type with a cloned Analysis and borrowed entry sets.