Type Alias rustc_mir_dataflow::ResultsRefCursor

pub type ResultsRefCursor<'res, 'mir, 'tcx, A> = ResultsCursor<'mir, 'tcx, A, &'res mut Results<'tcx, A>>;

A ResultsCursor that borrows the underlying Results.

Aliased Type

struct ResultsRefCursor<'res, 'mir, 'tcx, A> {
    body: &'mir Body<'tcx>,
    results: &'res mut Results<'tcx, A, IndexVec<BasicBlock, <A as AnalysisDomain<'tcx>>::Domain>>,
    state: <A as AnalysisDomain<'tcx>>::Domain,
    pos: CursorPosition,
    state_needs_reset: bool,
    reachable_blocks: BitSet<BasicBlock>,
}

Fields
body: &'mir Body<'tcx>results: &'res mut Results<'tcx, A, IndexVec<BasicBlock, <A as AnalysisDomain<'tcx>>::Domain>>state: <A as AnalysisDomain<'tcx>>::Domainpos: CursorPositionstate_needs_reset: bool

Indicates that state has been modified with a custom effect.

When this flag is set, we need to reset to an entry set before doing a seek.

reachable_blocks: BitSet<BasicBlock>
Layout

Note: Encountered an error during type layout; the type failed to be normalized.