rustc_mir_dataflow::value_analysis
Struct Map

pub struct Map<'tcx> {
    locals: IndexVec<Local, Option<PlaceIndex>>,
    projections: FxHashMap<(PlaceIndex, TrackElem), PlaceIndex>,
    places: IndexVec<PlaceIndex, PlaceInfo<'tcx>>,
    value_count: usize,
    inner_values: IndexVec<PlaceIndex, Range<usize>>,
    inner_values_buffer: Vec<ValueIndex>,
}

Partial mapping from Place to PlaceIndex, where some places also have a ValueIndex.

This data structure essentially maintains a tree of places and their projections. Some additional bookkeeping is done, to speed up traversal over this tree:

For iteration, every PlaceInfo contains an intrusive linked list of its children.

To directly get the child for a specific projection, there is a projections map.

Fields
locals: IndexVec<Local, Option<PlaceIndex>>projections: FxHashMap<(PlaceIndex, TrackElem), PlaceIndex>places: IndexVec<PlaceIndex, PlaceInfo<'tcx>>value_count: usizeinner_values: IndexVec<PlaceIndex, Range<usize>>inner_values_buffer: Vec<ValueIndex>
Implementations

impl<'tcx> Map<'tcx>

pub fn new( tcx: TyCtxt<'tcx>, body: &Body<'tcx>, value_limit: Option<usize>, ) -> Self

Returns a map that only tracks places whose type has scalar layout.

This is currently the only way to create a Map. The way in which the tracked places are chosen is an implementation detail and may not be relied upon (other than that their type are scalars).

fn register( &mut self, tcx: TyCtxt<'tcx>, body: &Body<'tcx>, exclude: BitSet<Local>, value_limit: Option<usize>, )

Register all non-excluded places that have scalar layout.

fn register_place( &mut self, ty: Ty<'tcx>, base: PlaceIndex, elem: TrackElem, ) -> PlaceIndex

fn cache_preorder_invoke(&mut self, root: PlaceIndex)

Precompute the list of values inside root and store it inside as a slice within inner_values_buffer.

impl<'tcx> Map<'tcx>

pub fn apply(&self, place: PlaceIndex, elem: TrackElem) -> Option<PlaceIndex>

Applies a single projection element, yielding the corresponding child.

fn find_extra( &self, place: PlaceRef<'_>, extra: impl IntoIterator<Item = TrackElem>, ) -> Option<PlaceIndex>

Locates the given place, if it exists in the tree.

pub fn find(&self, place: PlaceRef<'_>) -> Option<PlaceIndex>

Locates the given place, if it exists in the tree.

pub fn find_discr(&self, place: PlaceRef<'_>) -> Option<PlaceIndex>

Locates the given place and applies Discriminant, if it exists in the tree.

pub fn find_len(&self, place: PlaceRef<'_>) -> Option<PlaceIndex>

Locates the given place and applies DerefLen, if it exists in the tree.

fn children( &self, parent: PlaceIndex, ) -> impl Iterator<Item = PlaceIndex> + Captures<'_> + Captures<'tcx>

Iterate over all direct children.

fn for_each_aliasing_place( &self, place: PlaceRef<'_>, tail_elem: Option<TrackElem>, f: &mut impl FnMut(ValueIndex), )

Invoke a function on the given place and all places that may alias it.

In particular, when the given place has a variant downcast, we invoke the function on all the other variants.

tail_elem allows to support discriminants that are not a place in MIR, but that we track as such.

fn for_each_variant_sibling( &self, parent: PlaceIndex, preserved_child: Option<PlaceIndex>, f: &mut impl FnMut(ValueIndex), )

Invoke the given function on all the descendants of the given place, except one branch.

fn for_each_value_inside( &self, root: PlaceIndex, f: &mut impl FnMut(ValueIndex), )

Invoke a function on each value in the given place and all descendants.

pub fn for_each_projection_value<O>( &self, root: PlaceIndex, value: O, project: &mut impl FnMut(TrackElem, &O) -> Option<O>, f: &mut impl FnMut(PlaceIndex, &O), )

Invoke a function on each value in the given place and all descendants.