Crate rustc_mir_dataflow

Re-exports

pub use self::drop_flag_effects::drop_flag_effects_for_function_entry;

pub use self::drop_flag_effects::drop_flag_effects_for_location;

pub use self::drop_flag_effects::move_path_children_matching;

pub use self::drop_flag_effects::on_all_children_bits;

pub use self::drop_flag_effects::on_lookup_result_bits;

Modules

debuginfo

drop_flag_effects

elaborate_drops

errors 🔒

fmt

Custom formatting traits used when outputting Graphviz diagrams with the results of a dataflow analysis.

framework 🔒

A framework that can express both gen-kill and generic dataflow problems.

graphviz

A helpful diagram for debugging dataflow problems.

impls

Dataflow analyses are built upon some interpretation of the bitvectors attached to each basic block, represented via a zero-sized structure.

lattice

Traits used to represent lattices for use as the domain of a dataflow analysis.

move_paths

points

rustc_peek

storage

un_derefer

value_analysis

This module provides a framework on top of the normal MIR dataflow framework to simplify the implementation of analyses that track information about the values stored in certain places. We are using the term “place” here to refer to a mir::Place (a place expression) instead of an interpret::Place (a memory location).

Structs

Backward

Dataflow that runs from the exit of a block (the terminator), to its entry (the first statement).

Engine

A solver for dataflow problems.

Forward

Dataflow that runs from the entry of a block (the first statement), to its exit (terminator).

MoveDataParamEnv

Results

A dataflow analysis that has converged to fixpoint.

ResultsCursor

Allows random access inspection of the results of a dataflow analysis.

Enums

MaybeReachable

Extend a lattice with a bottom value to represent an unreachable execution.

Statics

DEFAULT_LOCALE_RESOURCE

Raw content of Fluent resource for this crate, generated by fluent_messages macro, imported by rustc_driver to include all crates’ resources in one bundle.

Traits

Analysis

A dataflow problem with an arbitrarily complex transfer function.

AnalysisDomain

Defines the domain of a dataflow problem.

Direction

GenKill

The legal operations for a transfer function in a gen/kill problem.

GenKillAnalysis

A gen/kill dataflow problem.

JoinSemiLattice

A partially ordered set that has a least upper bound for any pair of elements in the set.

ResultsVisitable

Things that can be visited by a ResultsVisitor.

ResultsVisitor

A visitor over the results of an Analysis. The type parameter R is the results type being visited.

SwitchIntEdgeEffects

A type that records the edge-specific effects for a SwitchInt terminator.

Functions

visit_results

Calls the corresponding method in ResultsVisitor for every location in a mir::Body with the dataflow state at that location.