Trait rustc_mir_dataflow::fmt::DebugWithContext

pub trait DebugWithContext<C>: Eq + Debug {
    fn fmt_with(&self, _ctxt: &C, f: &mut Formatter<'_>) -> Result { ... }
    fn fmt_diff_with(&self, old: &Self, ctxt: &C, f: &mut Formatter<'_>) -> Result { ... }
}

An extension to fmt::Debug for data that can be better printed with some auxiliary data C.

Provided Methods

fn fmt_with(&self, _ctxt: &C, f: &mut Formatter<'_>) -> Result

fn fmt_diff_with(&self, old: &Self, ctxt: &C, f: &mut Formatter<'_>) -> Result

Print the difference between self and old.

This should print nothing if self == old.

+ and - are typically used to indicate differences. However, these characters are fairly common and may be needed to print a types representation. If using them to indicate a diff, prefix them with the “Unit Separator” control character (␟ U+001F).