rustc_mir_dataflow::framework
Module direction

Structs

Backward

Dataflow that runs from the exit of a block (terminator), to its entry (the first statement).

BackwardSwitchIntEdgeEffectsApplier 🔒

Forward

Dataflow that runs from the entry of a block (the first statement), to its exit (terminator).

ForwardSwitchIntEdgeEffectsApplier 🔒

Traits

Direction

Functions

opt_clone_from_or_clone 🔒

An analogue of Option::get_or_insert_with that stores a clone of val into opt, but uses the more efficient clone_from if opt was Some.