rustc_mir_dataflow::impls::storage_liveness
Type Alias BorrowedLocalsResults

type BorrowedLocalsResults<'mir, 'tcx> = ResultsCursor<'mir, 'tcx, MaybeBorrowedLocals>;

Aliased Type

struct BorrowedLocalsResults<'mir, 'tcx> {
    body: &'mir Body<'tcx>,
    results: Results<'tcx, MaybeBorrowedLocals>,
    state: BitSet<Local>,
    pos: CursorPosition,
    state_needs_reset: bool,
    reachable_blocks: BitSet<BasicBlock>,
}

Fields
body: &'mir Body<'tcx>results: Results<'tcx, MaybeBorrowedLocals>state: BitSet<Local>pos: CursorPositionstate_needs_reset: bool

Indicates that state has been modified with a custom effect.

When this flag is set, we need to reset to an entry set before doing a seek.

reachable_blocks: BitSet<BasicBlock>
Layout

Note: Most layout information is completely unstable and may even differ between compilations. The only exception is types with certain repr(...) attributes. Please see the Rust Reference's “Type Layout” chapter for details on type layout guarantees.

Size: 128 bytes