Trait rustc_mir_dataflow::lattice::JoinSemiLattice

pub trait JoinSemiLattice: Eq {
    // Required method
    fn join(&mut self, other: &Self) -> bool;
}

A partially ordered set that has a least upper bound for any pair of elements in the set.

Required Methods

fn join(&mut self, other: &Self) -> bool

Computes the least upper bound of two elements, storing the result in self and returning true if self has changed.

The lattice join operator is abbreviated as ∨.

Object Safety

This trait is not object safe.