Trait rustc_mir_dataflow::lattice::MeetSemiLattice

pub trait MeetSemiLattice: Eq {
    // Required method
    fn meet(&mut self, other: &Self) -> bool;
}

A partially ordered set that has a greatest lower bound for any pair of elements in the set.

Dataflow analyses only require that their domains implement JoinSemiLattice, not MeetSemiLattice. However, types that will be used as dataflow domains should implement both so that they can be used with Dual.

Required Methods

fn meet(&mut self, other: &Self) -> bool

Computes the greatest lower bound of two elements, storing the result in self and returning true if self has changed.

The lattice meet operator is abbreviated as ∧.

Object Safety

This trait is not object safe.